% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_shadowtext.R
\name{element_shadowtext}
\alias{element_shadowtext}
\title{element_shadowtext}
\usage{
element_shadowtext(
  family = NULL,
  face = NULL,
  colour = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  color = NULL,
  margin = NULL,
  debug = NULL,
  bg.colour = "black",
  bg.color = NULL,
  bg.r = 0.1,
  inherit.blank = FALSE
)
}
\arguments{
\item{family}{Font family}

\item{face}{Font face ("plain", "italic", "bold", "bold.italic")}

\item{colour}{text colour}

\item{size}{text size in pts}

\item{hjust}{horizontal justification (in [0, 1])}

\item{vjust}{vertical justification (in [0, 1])}

\item{angle}{text angle}

\item{lineheight}{line height}

\item{color}{aliase to colour}

\item{margin}{margins around the text, see also 'margin()' for more details}

\item{debug}{if 'TRUE', aids visual debugging by drawing a solic rectangle behind the complete text area, and a point where each label is anchored.}

\item{bg.colour}{the background colour of text, default is \code{black}.}

\item{bg.color}{the alias of bg.colour,}

\item{bg.r}{background ratio of shadow text.}

\item{inherit.blank}{whether inherit 'element_blank'}
}
\value{
element_shadowtext object
}
\description{
shadown text element for ggplot theme system
}
\author{
Guangchuang Yu and xmarti6@github
}
