% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interactive.R
\name{geom_shadowtext_interactive}
\alias{geom_shadowtext_interactive}
\title{Create interactive shadow text}
\usage{
geom_shadowtext_interactive(...)
}
\arguments{
\item{...}{see also the parameters of `geom_shadowtext()` of `shadowtext`}
}
\description{
The geometry is based on `geom_shadowtext()`.
See the documentation for those functions for more details.
}
\examples{
library(ggplot2)
library(ggiraph)
mtcars$label <- rownames(mtcars)
mtcars$tooltip = paste0(
   "cyl: ",
   mtcars$cyl,
   "<br/>",
   "gear: ",
   mtcars$gear,
   "<br/>",
   "carb: ",
   mtcars$carb
 )
p <- ggplot(
       mtcars,
       aes(
         x = mpg,
         y = wt,
         label = label,
         color = qsec,
         tooltip = tooltip,
         data_id = label
       )
     ) +
     geom_shadowtext_interactive(check_overlap = TRUE) +
     coord_cartesian(xlim = c(0, 50)) 
girafe(ggobj = p, options = list(opts_hover(css = "fill:#FF4C3B;font-style:italic;")))
}
