% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-shadowtext.R
\docType{data}
\name{GeomShadowText}
\alias{GeomShadowText}
\alias{GeomShadowtext}
\title{GeomShadowText}
\format{
An object of class \code{GeomShadowText} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomShadowText
}
\description{
ShadowText Geom
}
\author{
Guangchuang Yu
}
\keyword{datasets}
