% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtext-grob.R
\name{shadowtextGrob}
\alias{shadowtextGrob}
\alias{grid.shadowtext}
\title{shadowtextGrob}
\usage{
shadowtextGrob(label, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  just = "centre", hjust = NULL, vjust = NULL, rot = 0,
  check.overlap = FALSE, default.units = "npc", name = NULL,
  gp = gpar(col = "white"), vp = NULL, bg.color = "black", bg.r = 0.1)

grid.shadowtext(label, x = unit(0.5, "npc"), y = unit(0.5, "npc"),
  just = "centre", hjust = NULL, vjust = NULL, rot = 0,
  check.overlap = FALSE, default.units = "npc", name = NULL,
  gp = gpar(col = "white"), vp = NULL, bg.color = "black", bg.r = 0.1,
  draw = TRUE)
}
\arguments{
\item{label}{text labels}

\item{x}{x position}

\item{y}{y position}

\item{just}{The justification of the text, can be 'left', 'right', 'center', 'bottom' and 'top'}

\item{hjust}{horizontal adjustment}

\item{vjust}{vertical adjustment}

\item{rot}{angle to rotate the text}

\item{check.overlap}{whether check for and omit overlapping text}

\item{default.units}{default unit of x and y}

\item{name}{identifier}

\item{gp}{gpar object}

\item{vp}{viewport object}

\item{bg.color}{background color of shadow text}

\item{bg.r}{background ratio of shadow text}

\item{draw}{whether draw the grob}
}
\value{
grob object
}
\description{
create and draw text
}
\author{
guangchuang yu
}
