% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interactive.R
\name{geom_shadowtext_interactive}
\alias{geom_shadowtext_interactive}
\title{Create interactive shadow text}
\usage{
geom_shadowtext_interactive(...)
}
\arguments{
\item{...}{see also the parameters of `geom_shadowtext()` of `shadowtext`}
}
\description{
The geometry is based on `geom_shadowtext()`.
See the documentation for those functions for more details.
}
