% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowtext-package.R
\docType{package}
\name{shadowtext-package}
\alias{shadowtext}
\alias{shadowtext-package}
\title{shadowtext: Shadow Text Grob and Layer}
\description{
Implement shadowtextGrob() for 'grid' and geom_shadowtext() layer for 'ggplot2'. These functions create/draw text grob with background shadow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GuangchuangYu/shadowtext/}
  \item Report bugs at \url{https://github.com/GuangchuangYu/shadowtext/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
}

}
\keyword{internal}
